+++ R/makeSummaryTTE.R
@@ -13,6 +13,7 @@
 #' @param showAvnObs a logical, show a column with the average number of events? Default will be `FALSE` if the data is recognized as RTTE data, `TRUE` if not.
 #' @param nObsColNm is the character string to be printed as the column name with the number of observations, default is "\\\\textbf\{nEvent\\\\textsuperscript\{b\}\}"
 #' @param avnObsColNm is the character string to be printed as the column name with the average number of observations per subject in a given strata, default is "\\\\textbf\{pEvent\\\\textsuperscript\{c\}\}"
+#' @param avnObsPercent is a logical, should "pEvent" be printed as a percentage? TRUE
 #' @param caption is the table caption. Assign NULL to this argument produce table without caption. Default is "Number of patients and number of events"
 #' @param footnote is the text for footnote, default is "\\\\textsuperscript\{a\}Number of subjects\\\\newline\\\\textsuperscript\{b\}Number of events\\\\newline\\\\textsuperscript\{c\}Proportion of number of events" for TTE data.
 #' @param myFun internal function for the calculation of summarized data. If NULL, the default, an internal function specific to R(TTE) data is used. This should not be changed for a standard use.
@@ -55,6 +56,7 @@ makeSummaryTableTTE <- function(df,
                                 showAvnObs = !isRTTE(df, myID = myID),
                                 nObsColNm = "\\textbf{nEvent\\textsuperscript{b}}",
                                 avnObsColNm = "\\textbf{pEvent\\textsuperscript{c}}",
+                                avnObsPercent = TRUE,
                                 caption = "Number of patients and number of events",
                                 footnote = paste0(
                                   "\\textsuperscript{a}Number of subjects\\newline",
@@ -70,12 +72,15 @@ makeSummaryTableTTE <- function(df,
         summarise(
           subjects = length(unique(!!rlang::sym(myID))),
           nObs     = length(c(!!rlang::sym(myDV))[!!rlang::sym(myDV)==1]),
-          avnObs   = PhRame_out.digits(
-            length(c(!!rlang::sym(myDV))[!!rlang::sym(myDV)==1])/length(unique(!!rlang::sym(myID))),
-            dig = digits,
-            numeric = F)
-        )
-      if(!showAvnObs){
+          avnObs   = length(c(!!rlang::sym(myDV))[!!rlang::sym(myDV)==1])/length(unique(!!rlang::sym(myID)))
+          )
+      if(showAvnObs){
+        if(avnObsPercent){
+          ans$avnObs <- scales::percent(ans$avnObs, accuracy = (10^2 * 10^-digits))
+        } else {
+          ans$avnObs <- PhRame_out.digits(ans$avnObs, dig = digits, numeric = FALSE)
+        }
+      } else {
         ans$avnObs <- ""
       }
       ans